+++ examples/authorizer/bun/authorizer.ts
@@ -1,7 +1,7 @@
-import { authorizer } from "@openauthjs/openauth"
-import { MemoryStorage } from "@openauthjs/openauth/storage/memory"
-import { PasswordAdapter } from "@openauthjs/openauth/adapter/password"
-import { PasswordUI } from "@openauthjs/openauth/ui/password"
+import { authorizer } from "@clopca/openauth"
+import { MemoryStorage } from "@clopca/openauth/storage/memory"
+import { PasswordAdapter } from "@clopca/openauth/adapter/password"
+import { PasswordUI } from "@clopca/openauth/ui/password"
 import { subjects } from "../../subjects.js"
 
 async function getUser(email: string) {