+++ examples/client/jwt-api/index.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@openauthjs/openauth/client"
+import { createClient } from "@clopca/openauth/client"
 import { subjects } from "../../subjects"
 
 const headers = {